Move app main logic to Program class, temporary main app logit for testing
